More Compare and AutoTests to respective classes
