+++ UnitTests/CombinedUnits/AccelerationTests.cs
@@ -11,6 +11,57 @@ namespace UnitTests.CombinedUnits;
 public class AccelerationTests
 {
     [TestMethod]
+    public void AutoTest()
+    {
+        var A1 = new UnitsNet.Acceleration(65.743, UnitsNet.Units.AccelerationUnit.MeterPerSecondSquared);
+        var A2 = new EngineeringUnits.Acceleration(65.743, AccelerationUnit.MeterPerSecondSquared);
+
+        var WorkingCompares = 0;
+
+        foreach (AccelerationUnit EU in UnitTypebase.ListOf<AccelerationUnit>())
+        {
+
+            var Error = 2E-4;
+            var RelError = 1E-5;
+
+            IEnumerable<UnitsNet.Units.AccelerationUnit> UNList = UnitsNet.Acceleration.Units.Where(x => x.ToString() == EU.QuantityName);
+
+            if (UNList.Count() == 1)
+            {
+                UnitsNet.Units.AccelerationUnit UN = UNList.Single();
+
+                //if (UN == UnitsNet.Units.AccelerationUnit.SquareMicrometer) Error = 2629720.0009765625;
+
+                Debug.Print($"");
+                Debug.Print($"UnitsNets:       {UN} {A1.As(UN)}");
+                Debug.Print($"EngineeringUnit: {EU.QuantityName} {A2.As(EU)}");
+                Debug.Print($"ABS:    {A2.As(EU) - A1.As(UN):F6}");
+                Debug.Print($"REF[%]: {HelperClass.Percent(A2.As(EU), A1.As(UN)):P6}");
+
+                //All units absolute difference
+                Assert.AreEqual(0, A2.As(EU) - A1.As(UN), Error);
+
+                //All units relative difference
+                Assert.AreEqual(0, HelperClass.Percent(A2.As(EU),
+                                                        A1.As(UN)),
+                                                        RelError);
+                //All units symbol compare
+                Assert.AreEqual(A2.ToUnit(EU).DisplaySymbol(),
+                                A1.ToUnit(UN).ToString("a")
+                                //.Replace("min", "m")
+                                );
+
+                WorkingCompares++;
+
+            }
+        }
+
+        //Number of comparables units
+        Assert.AreEqual(14, WorkingCompares);
+
+    }
+
+    [TestMethod]
     public void Adding()
     {
         var A1 = new UnitsNet.Acceleration(1, UnitsNet.Units.AccelerationUnit.FootPerSecondSquared);
