+++ UnitTests/CombinedUnits/DensityTests.cs
@@ -2,6 +2,7 @@ using EngineeringUnits;
 using EngineeringUnits.Units;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Collections.Generic;
+using System.Diagnostics;
 using System.Linq;
 
 namespace UnitTests.CombinedUnits;
@@ -9,6 +10,58 @@ namespace UnitTests.CombinedUnits;
 [TestClass]
 public class DensityTests
 {
+
+    [TestMethod]
+    public void AutoTest()
+    {
+        var A1 = new UnitsNet.Density(1, UnitsNet.Units.DensityUnit.KilogramPerCubicMeter);
+        var A2 = new EngineeringUnits.Density(1, DensityUnit.KilogramPerCubicMeter);
+
+        var WorkingCompares = 0;
+
+        foreach (DensityUnit EU in UnitTypebase.ListOf<DensityUnit>())
+        {
+
+            var Error = 1E-5;
+            var RelError = 2E-4;
+
+            IEnumerable<UnitsNet.Units.DensityUnit> UNList = UnitsNet.Density.Units.Where(x => x.ToString() == EU.QuantityName);
+
+            if (UNList.Count() == 1)
+            {
+                UnitsNet.Units.DensityUnit UN = UNList.Single();
+
+                //if (UN == UnitsNet.Units.DensityUnit.SquareMicrometer) Error = 2629720.0009765625;
+
+                Debug.Print($"");
+                Debug.Print($"UnitsNets:       {UN} {A1.As(UN)}");
+                Debug.Print($"EngineeringUnit: {EU.QuantityName} {A2.As(EU)}");
+                Debug.Print($"ABS:    {A2.As(EU) - A1.As(UN):F6}");
+                Debug.Print($"REF[%]: {HelperClass.Percent(A2.As(EU), A1.As(UN)):P6}");
+
+                //All units absolute difference
+                Assert.AreEqual(0, A2.As(EU) - A1.As(UN), Error);
+
+                //All units relative difference
+                Assert.AreEqual(0, HelperClass.Percent(A2.As(EU),
+                                                        A1.As(UN)),
+                                                        RelError);
+                //All units symbol compare
+                Assert.AreEqual(A2.ToUnit(EU).DisplaySymbol(),
+                                A1.ToUnit(UN).ToString("a")
+                                .Replace("/L", "/l")
+                                //.Replace("kip", "klb")
+                                );
+
+                WorkingCompares++;
+
+            }
+        }
+
+        //Number of comparables units
+        Assert.AreEqual(40, WorkingCompares);
+
+    }
 
     [TestMethod]
     public void Adding()
