+++ UnitTests/CombinedUnits/PressureTests.cs
@@ -1,12 +1,68 @@
 using EngineeringUnits;
 using EngineeringUnits.Units;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
 
 namespace UnitTests.CombinedUnits;
 
 [TestClass]
 public class PressureTests
 {
+
+    [TestMethod]
+    public void AutoTest()
+    {
+        var A1 = new UnitsNet.Pressure(65.743, UnitsNet.Units.PressureUnit.Bar);
+        var A2 = new EngineeringUnits.Pressure(65.743, PressureUnit.Bar);
+
+        var WorkingCompares = 0;
+
+        foreach (PressureUnit EU in UnitTypebase.ListOf<PressureUnit>())
+        {
+
+            var Error = 1E-5;
+            var RelError = 1E-5;
+
+            IEnumerable<UnitsNet.Units.PressureUnit> UNList = UnitsNet.Pressure.Units.Where(x => x.ToString() == EU.QuantityName);
+
+            if (UNList.Count() == 1)
+            {
+                UnitsNet.Units.PressureUnit UN = UNList.Single();
+
+                if (UN == UnitsNet.Units.PressureUnit.Micropascal)
+                    Error = 0.0009765625;
+                if (UN == UnitsNet.Units.PressureUnit.KilogramForcePerSquareMeter)
+                    Error = 0.013645294005982578;
+
+                Debug.Print($"");
+                Debug.Print($"UnitsNets:       {UN} {A1.As(UN)}");
+                Debug.Print($"EngineeringUnit: {EU.QuantityName} {A2.As(EU)}");
+                Debug.Print($"ABS:    {A2.As(EU) - A1.As(UN):F6}");
+                Debug.Print($"REF[%]: {HelperClass.Percent(A2.As(EU), A1.As(UN)):P6}");
+                Debug.Print($"UnitsNets:       {A1.ToUnit(UN):a}");
+                Debug.Print($"EngineeringUnit: {A2.ToUnit(EU).DisplaySymbol()}");
+
+                //All units absolute difference
+                Assert.AreEqual(0, A2.As(EU) - A1.As(UN), Error);
+
+                //All units relative difference
+                Assert.AreEqual(0, HelperClass.Percent(A2.As(EU),
+                                                        A1.As(UN)),
+                                                        RelError);
+                //All units symbol compare
+                Assert.AreEqual(A2.ToUnit(EU).DisplaySymbol(),
+                                A1.ToUnit(UN).ToString("a"));
+
+                WorkingCompares++;
+            }
+        }
+
+        //Number of comparables units
+        Assert.AreEqual(42, WorkingCompares);
+    }
+
     [TestMethod]
     public void PressureEquals()
     {
