+++ UnitTests/CombinedUnits/SpeedTests.cs
@@ -2,12 +2,67 @@ using EngineeringUnits;
 using EngineeringUnits.Units;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
 
 namespace UnitTests.CombinedUnits;
 
 [TestClass]
 public class SpeedTests
 {
+
+    [TestMethod]
+    public void AutoTest()
+    {
+        var A1 = new UnitsNet.Speed(65.743, UnitsNet.Units.SpeedUnit.FootPerHour);
+        var A2 = new EngineeringUnits.Speed(65.743, SpeedUnit.FootPerHour);
+
+        var WorkingCompares = 0;
+
+        foreach (SpeedUnit EU in UnitTypebase.ListOf<SpeedUnit>())
+        {
+
+            var Error = 1E-5;
+            var RelError = 1E-4;
+
+            IEnumerable<UnitsNet.Units.SpeedUnit> UNList = UnitsNet.Speed.Units.Where(x => x.ToString() == EU.QuantityName);
+
+            if (UNList.Count() == 1)
+            {
+                UnitsNet.Units.SpeedUnit UN = UNList.Single();
+
+                //if (UN == UnitsNet.Units.SpeedUnit.SquareMicrometer) Error = 2629720.0009765625;
+
+                Debug.Print($"");
+                Debug.Print($"UnitsNets:       {UN} {A1.As(UN)}");
+                Debug.Print($"EngineeringUnit: {EU.QuantityName} {A2.As(EU)}");
+                Debug.Print($"ABS:    {A2.As(EU) - A1.As(UN):F6}");
+                Debug.Print($"REF[%]: {HelperClass.Percent(A2.As(EU), A1.As(UN)):P6}");
+
+                //All units absolute difference
+                Assert.AreEqual(0, A2.As(EU) - A1.As(UN), Error);
+
+                //All units relative difference
+                Assert.AreEqual(0, HelperClass.Percent(A2.As(EU),
+                                                        A1.As(UN)),
+                                                        RelError);
+                //All units symbol compare
+                Assert.AreEqual(A2.ToUnit(EU).DisplaySymbol(),
+                                A1.ToUnit(UN).ToString("a")
+                                //.Replace("/min", "/m")
+                                );
+
+                WorkingCompares++;
+
+            }
+        }
+
+        //Number of comparables units
+        Assert.AreEqual(32, WorkingCompares);
+
+    }
+
     [TestMethod]
     public void Simple()
     {
